requirements.txt
Flask==3.0.2
openai==1.14.0
pandas==2.2.1
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-Limiter==3.5.0
